move helper to core-utils
