+++ packages/utils/core-utils/src/index.ts
@@ -1,3 +1,4 @@
+export * from './hook.js';
 export * from './rebuild.js';
 export * from './electron-version.js';
 export * from './package-manager.js';
